test: split interaction runtime tests
